# src/main/java/com/webservice/weather/model/Forecast.java
package com.webservice.weather.model;

import java.time.LocalDate;

public class Forecast {

	LocalDate date;
	String avgTemp;
	String weather;

	public LocalDate getDate() {
		return date;
	}

	public void setDate(LocalDate date) {
		this.date = date;
	}

	public String getAvgTemp() {
		return avgTemp;
	}

	public void setAvgTemp(String avgTemp) {
		this.avgTemp = avgTemp;
	}

	public String getWeather() {
		return weather;
	}

	public void setWeather(String weather) {
		this.weather = weather;
	}
}