# src/main/resources/api-spec.yaml
swagger: "2.0"
info:
  description: "This is the project for getting the current and forecast weather information"
  version: "1.0.0"
  title: "Weather API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "mandovarakrishna@gmail.com"
host: "localhost:8080"
basePath: "/api"
tags:
- name: "Weather"
  description: "Weather Information - Current and Forecast"
  externalDocs:
    description: "Find out more"
    url: "http://swagger.io"
schemes:
- "http"
paths:
  /current-weather:
    get:
      tags:
      - "Weather"
      summary: "Get Current weather information"
      description: "Information can obtained by city or zipCode"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
      - name: "city"
        in: "query"
        description: "Weather Information for the city"
        required: false
        type: "string"
      - name: "zipCode"
        in: "query"
        description: "Weather Information for the zip-code"
        required: false
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/CurrentWeatherInfo"
        "400":
          description: "Invalid status value"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Error"
        "401":
          description: "Invalid status value"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UnAuthorizedError"
        "404":
          description: "Invalid status value"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Error"
      security:
      - api_key: []
        
  /forecast-weather:
    get:
      tags:
      - "Weather"
      summary: "Get Forecast weather information"
      description: "Information can obtained by city or zipCode"
      produces:
      - "application/json"
      parameters:
      - name: "city"
        in: "query"
        description: "Weather Information for the city"
        required: false
        type: "string"
      - name: "zipCode"
        in: "query"
        description: "Weather Information for the zip-code"
        required: false
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ForecastWeatherInfo"
        "400":
          description: "Invalid status value"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Error"
        "401":
          description: "Invalid status value"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UnAuthorizedError"
        "404":
          description: "Invalid status value"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Error"
      security:
      - api_key: []

securityDefinitions:
  api_key:
    type: "apiKey"
    name: "apiKey"
    in: "query"
    
definitions:

  ForecastWeatherInfo:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      forecast:
        type: "array"
        items:
          $ref: "#/definitions/Forecast"
        description: "Aggregated temperature from different API, valid till two decimal places in Celsius"
      city:
        type: "string"
        description: "Current weather information of this city"
        example: "Boston"
        
  Forecast:
    type: "object"
    properties:
      date:
        type: "string"
        description: "Forcast date"
        example: "2020-12-03"
        format: date
      avgTemp:
        type: "string"
        description: "Aggregated temperature from different API, valid till two decimal places in Celsius"
        example: "29.07"
      weather:
        type: "string"
        description: "Forecast weather condition of the city on the date"
        example: "Clear sky"
      
  CurrentWeatherInfo:
    type: "object"
    properties:
      avgTemp:
        type: "string"
        description: "Aggregated temperature from different API, valid till two decimal places in Celsius"
        example: "29.07"
      avgApparent:
        type: "string"
        description: "Aggregated feel-like temprature from different API, valid till two decimal places in Celsius"
        example: "27.67"
      city:
        type: "string"
        description: "Current weather information of this city"
        example: "Boston"
      weather:
        type: "string"
        description: "Current weather condition of the city"
        example: "Clear sky"
      date:
        type: "string"
        format: "date"
        description: "Current Date"
        example: "2020-12-03"
        
  Error:
    type: "object"
    properties:
      message:
        type: "string"
        description: "Message for the error occured"
        example: "Error while processing the request"
      type:
        type: "string"
        description: "Type for the error occured"
        example: "SERVICE_UNAVAILABLE"
        
  UnAuthorizedError:
    type: "object"
    properties:
      message:
        type: "string"
        description: "Message for the error occured"
        example: "Invalid API key."
      type:
        type: "string"
        description: "Type for the error occured"
        example: "Unauthorized"
    
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"